refactor(db): extract SFX functions to src/db/sfx.ts (PR 8)
src/db.ts is now a pure barrel — all logic lives in focused modules under src/db/.
